Moved all componenets to a component folder
