+++ frontend/src/components/Sliders/BrightnessSlider.jsx
