+++ frontend/src/components/Sliders/ColorSlider.jsx
