+++ frontend/src/pages/Lights.jsx
@@ -1,6 +1,6 @@
-import HeaderBar from "../HeaderBar/HeaderBar"
-import ColorSlider from "../Sliders/ColorSlider"
-import BrightnessSlider from "../Sliders/BrightnessSlider"
+import HeaderBar from "../components/HeaderBar/HeaderBar"
+import ColorSlider from "../components/Sliders/ColorSlider"
+import BrightnessSlider from "../components/Sliders/BrightnessSlider"
 
 const Lights = () => {
   return (
